pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dorothy"
version = "0.1.0"
description = "Deep learning framework for inferring stellar parameters from spectroscopic data"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "DOROTHY Team"}
]
keywords = ["astronomy", "spectroscopy", "deep-learning", "stellar-parameters"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "astropy>=5.0.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
dorothy = "dorothy.cli.main:main"

[project.urls]
Homepage = "https://github.com/dorothy-project/dorothy"
Documentation = "https://github.com/dorothy-project/dorothy#readme"
Repository = "https://github.com/dorothy-project/dorothy"

[tool.setuptools.packages.find]
where = ["."]
include = ["dorothy*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line length handled separately

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
